Make the progress bar argument a double instead of an int.  This gets around an issue with guile 1.6.
git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@7832 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ src/app-file/gnc-file.h
@@ -134,7 +134,7 @@ void gnc_file_quit (void);
 typedef gboolean (*GNCCanCancelSaveCB) (void);
 void gnc_file_set_can_cancel_callback (GNCCanCancelSaveCB cb);
 
-typedef void (*GNCFilePercentageFunc) (const char *message, int percent);
+typedef void (*GNCFilePercentageFunc) (const char *message, double percent);
 void gnc_file_set_pct_handler (GNCFilePercentageFunc file_percentage_func);
 
 typedef void (*GNCShutdownCB) (int);
